fix queryset after move
