+++ edgy/core/db/querysets/mixins/queryset_props.py
@@ -8,10 +8,12 @@ import sqlalchemy
 from edgy.core.connection.database import Database
 
 if TYPE_CHECKING:
-    from edgy import QuerySet
+    from edgy.core.db.querysets.queryset import QuerySet as QuerySetAsType
+else:
+    QuerySetAsType = object
 
 
-class QuerySetPropsMixin:
+class QuerySetPropsMixin(QuerySetAsType):
     """
     A mixin class providing essential properties for QuerySet objects.
 
@@ -47,7 +49,7 @@ class QuerySetPropsMixin:
         return self._database
 
     @database.setter
-    def database(self: QuerySet, value: Database) -> None:
+    def database(self, value: Database) -> None:
         """
         Sets the database instance for the queryset and clears any cached table
         definitions.
@@ -83,7 +85,7 @@ class QuerySetPropsMixin:
         return self._table
 
     @table.setter
-    def table(self: QuerySet, value: sqlalchemy.Table | None) -> None:
+    def table(self, value: sqlalchemy.Table | None) -> None:
         """
         Sets the SQLAlchemy Table object for the queryset and clears the cache.
